initial cleanup from main
