--- experiments/experimental_src/helper.py
@@ -1,12 +1,0 @@
-import time # timing tests
-
-# helper function to test timings
-def timeit(func):
-    def wrapper(*args, **kwargs):
-        start_time = time.time()
-        result = func(*args, **kwargs)
-        end_time = time.time()
-        run_time = end_time - start_time
-        print(f"Function {func.__name__} took {run_time:.6f} seconds to run.")
-        return result
-    return wrapper
