+++ requirements-client.txt
@@ -1,0 +1,3 @@
+pyalsaaudio
+python-dotenv
+
