+++ requirements-server.txt
@@ -3,5 +3,6 @@ SpeechRecognition
 numpy
 pygame
 pydub
-pyalsaaudio
 ffmpeg-python
+python-dotenv
+openai
